% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_from_bucket.R
\name{cp_from_bucket}
\alias{cp_from_bucket}
\title{Copy a file from the bucket to your local workspace}
\usage{
cp_from_bucket(from, to)
}
\arguments{
\item{from}{the bucket file to be copied}

\item{to}{the destination on your local workspace}
}
\description{
Copy a file from the bucket to your local workspace
}
\examples{
\dontrun{
write.csv(c(1,2,3),"mydata.csv")
cp_to_bucket("mydata.csv","mydata.csv")
file.remove("mydata.csv")
list.files() # mydata.csv is not listed
cp_from_bucket("mydata.csv",".")
list.files() # mydata.csv has been copied
}
}
