% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_to_bucket.R
\name{cp_to_bucket}
\alias{cp_to_bucket}
\title{Copy a file to the bucket}
\usage{
cp_to_bucket(from, to)
}
\arguments{
\item{from}{the file in your local workspace to be copied}

\item{to}{the bucket destination}
}
\description{
Copy a file to the bucket
}
\examples{
\dontrun{
write.csv(c(1,2,3),"mydata.csv")
cp_to_bucket("mydata.csv","mydata.csv")
dat <- read_bucket("mydata.csv")
}
}
