% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bucket.R
\name{read_bucket}
\alias{read_bucket}
\title{Read bucket}
\usage{
read_bucket(file)
}
\arguments{
\item{file}{the file name}
}
\value{
a data.table
}
\description{
Read bucket
}
\examples{
\dontrun{
dat <- read_bucket("my_data.csv")
}
}
