% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_bucket.R
\name{rm_bucket}
\alias{rm_bucket}
\title{Remove a file from the workspace bucket}
\usage{
rm_bucket(path)
}
\arguments{
\item{path}{the path to the file to be deleted}
}
\description{
Remove a file from the workspace bucket
}
\examples{
\dontrun{
write.csv(c(1,2,3),"mydata.csv")
cp_to_bucket("mydata.csv","mydata.csv")
rm_bucket("mydata.csv")

# Remove file from arbitrary bucket
rm_bucket("gs://fc-secure-123/myfile.csv")
}
}
