% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_bucket.R
\name{ls_bucket}
\alias{ls_bucket}
\title{List the files in a bucket directory}
\usage{
ls_bucket(folder = NULL)
}
\arguments{
\item{folder}{the name of the folder}
}
\description{
List the files in a bucket directory
}
\examples{
\dontrun{
ls_bucket() # Files in the current workspace directory
ls_bucket("notebooks") # Files in the "notebooks" workspace directory
ls_bucket("gs://fc-secure-1234") # Files in another workspace
}
}
